Move util_cast > util/cast
