+++ libmamba/src/core/pool.cpp
@@ -20,8 +20,8 @@ extern "C"  // Incomplete header
 #include "mamba/core/context.hpp"
 #include "mamba/core/output.hpp"
 #include "mamba/core/pool.hpp"
-#include "mamba/core/util_cast.hpp"
 #include "mamba/core/util_string.hpp"
+#include "mamba/util/cast.hpp"
 #include "mamba/util/compare.hpp"
 #include "solv-cpp/queue.hpp"
